+++ starlite/utils/__init__.py
@@ -1,17 +1,18 @@
 # flake8: noqa
+from starlite.signature import (
+    SignatureModel,
+    get_signature_model,
+    model_function_signature,
+)
+
 from .model import convert_dataclass_to_model, create_parsed_model_field
 from .sequence import find_index, unique
-from .signature import (
-    SignatureModel,
-    create_function_signature_model,
-    get_signature_model,
-)
 from .url import join_paths, normalize_path
 
 __all__ = [
     "SignatureModel",
     "convert_dataclass_to_model",
-    "create_function_signature_model",
+    "model_function_signature",
     "create_parsed_model_field",
     "find_index",
     "get_signature_model",