+++ tests/openapi/test_path_item.py
@@ -1,13 +1,13 @@
-from starlite import Router
+from starlite import Starlite
 from starlite.openapi.path_item import create_path_item
 from starlite.utils import find_index
 from tests.openapi.utils import PersonController
 
 
 def test_create_path_item():
-    router = Router(path="", route_handlers=[PersonController])
-    index = find_index(router.routes, lambda x: x.path_format == "/{service_id}/person/{person_id}")
-    route = router.routes[index]
+    app = Starlite(route_handlers=[PersonController], openapi_config=None)
+    index = find_index(app.routes, lambda x: x.path_format == "/{service_id}/person/{person_id}")
+    route = app.routes[index]
     schema = create_path_item(route=route, create_examples=True)
     assert schema.delete
     assert schema.delete.operationId == "delete_person"